Refactor the image resize worker
We want to inject the aiosession into the consumer with all of the other IO objects so it can be easily substituted for testing purposes. Move the miscellaneous little supporting functions that aren't directly related to consumption to their own module.
