+++ image_get/worker/dummy_producer.py
@@ -22,5 +22,5 @@ encoded_msgs = [json.dumps(msg) for msg in msgs]
 client = KafkaClient(hosts='kafka:9092')
 topic = client.topics['inbound_images']
 with topic.get_sync_producer() as producer:
-    for url in islice(cycle(encoded_msgs), 2000):
+    for url in islice(cycle(encoded_msgs), 5000):
         producer.produce(bytes(url, 'utf-8'))
