+++ image_get/worker/util.py
@@ -1,0 +1,44 @@
+import json
+import logging as log
+import time
+from io import BytesIO
+
+import pykafka
+from PIL import Image
+
+import settings
+
+
+def kafka_connect():
+    try:
+        client = pykafka.KafkaClient(hosts=settings.KAFKA_HOSTS)
+    except pykafka.exceptions.NoBrokersAvailableError:
+        log.info('Retrying Kafka connection. . .')
+        time.sleep(3)
+        return kafka_connect()
+    return client
+
+
+def parse_message(message):
+    decoded = json.loads(str(message.value, 'utf-8'))
+    return decoded
+
+
+def thumbnail_image(img: Image):
+    img.thumbnail(size=settings.TARGET_RESOLUTION, resample=Image.NEAREST)
+    output = BytesIO()
+    img.save(output, format="JPEG", quality=30)
+    output.seek(0)
+    return output
+
+
+def save_thumbnail_s3(s3_client, img: BytesIO, identifier):
+    s3_client.put_object(
+        Bucket='cc-image-analysis',
+        Key=f'{identifier}.jpg',
+        Body=img
+    )
+
+
+def save_thumbnail_local(img: BytesIO):
+    pass
